implement font and contrast mods
